before: replace the borrowed bit view with the crate-owned BitsView
The walk surface reads stored streams and door buffers through
BitsView — borrowed bytes beside a u64 live bit length — so every
semantic walk (comparisons, joins' input side, admission walks, rank
folds, the id operations) is exact at every storable size on every
target. Bits's Deref to a borrowed bit slice, whose 32-bit length
encoding capped walks at 64 MiB, is deleted rather than widened;
Bits::live() is the explicit accessor in its place. Views are
byte-aligned by construction, so sub-stream ranges travel as explicit
(view, start, end) bit positions at the copy and parse seams.

The parallel byte-form duals existed only because the slice form
capped; each collapses to one form under the plain name:
require_marker_padding_bytes -> require_marker_padding,
parse_id_bytes -> parse_id, validate_prefix_bytes -> validate_prefix,
validate_dominating_bytes -> validate_dominating_from,
decode_int_window_bytes -> decode_int_window,
DsiCursor::over_bytes/over_bytes_live -> DsiCursor::new (BitsView::whole
/ BitsView::new carry the byte doors' forms), LeafCursor::open_bytes ->
LeafCursor::open. codec::byte_view dissolves into BitsView::body_tail —
views are byte-aligned by type, so its mid-byte None arm has no
representable input. The never-a-borrowed-bit-view door comments are
excised; prose is restated against the view.

The wasm32 pins flip in the same commit, per the pins file's own
red-first discipline: the composite-door pins (Ranked byte + borsh,
borsh Span) and the comparison pins at the 2^29-bit boundary go from
trap-assertions to correct-value assertions, with every adjacency
witness unchanged, and four new pins spot-check exactness deep in the
storable range (comparison at the 512 MiB storage bound; the composite
doors at 128-256 MiB components). The covered-join pins stay pinned as
traps with their mechanism re-attributed: the walk is exact at those
sizes, and the residual trap is the emitters' build-buffer length
encoding at the freeze seam — the same boundary the emitter-output pin
holds, cured by the crate-owned build buffer, not by this change. The
rank-cap and gap-boundary pins stay as found.

The skyline decode wrap now allocates its copy once, exactly sized plus
the marker bit (the meter envelopes tighten in the next commit). The
mutants roster's one line-pinned recode entry re-aims 457:38 -> 468:38
(same width-tie comparison, same equivalence argument; the counts
re-verify under just mutants-list).

+++ crates/before/src/party/ops/sum.rs
@@ -29,7 +29,7 @@ impl IdReader<'_> {
         // Conservative: the disjoint union has at most as many bits as both
         // inputs combined; normalization (collapsing `(v, v)` leaves) only
         // shrinks it. No tighter bound is cheap without doing the sum.
-        let mut out = IdBuilder::with_capacity(self.bits().len() + other.bits().len());
+        let mut out = IdBuilder::with_capacity((self.bits().len() + other.bits().len()) as usize);
         let mut frames = Frames::new();
         // Whether the current pair's side is a present child (read the real
         // cursor) or an absent `0` (stand in a synthetic empty).
